refactor: move catalogs to data manager and create a data factory

+++ MMORPGEngine/Data/Ground/GroundCatalog.cpp
@@ -3,9 +3,10 @@
 namespace Engine {
 
 GroundCatalog::GroundCatalog() :
-    _grounds() {}
+    _grounds() {
+}
 
-const GroundModel* GroundCatalog::ground( uint16_t type ) const {
+const GroundModel* GroundCatalog::ground( uint32_t type ) const {
     auto iterator = _grounds.find( type );
 
     if ( iterator == _grounds.end() ) {
@@ -15,7 +16,7 @@ const GroundModel* GroundCatalog::ground( uint16_t type ) const {
     return &iterator->second;
 }
 
-const std::unordered_map<uint16_t, GroundModel>& GroundCatalog::grounds() const {
+const std::unordered_map<uint32_t, GroundModel>& GroundCatalog::grounds() const {
     return _grounds;
 }
